back/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

# Banco de Dados e ORM
sqlmodel==0.0.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# Autenticação e Segurança
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# Tarefas Assíncronas
celery==5.4.0
redis==5.0.4

# Requisições HTTP (para APIs externas)
httpx==0.27.0

# Processamento de Arquivos
pdfplumber==0.11.0

# APIs do Google
google-cloud-vision==3.7.2
google-generativeai==0.7.1
google-auth==2.27.0

# Outras dependências necessárias
typing-extensions==4.12.2
python-multipart==0.0.9